SIG: Move sigcmn to internal
