+++ go/sig/BUILD.bazel
@@ -19,9 +19,9 @@ go_library(
         "//go/sig/internal/disp:go_default_library",
         "//go/sig/internal/ingress:go_default_library",
         "//go/sig/internal/metrics:go_default_library",
+        "//go/sig/internal/sigcmn:go_default_library",
         "//go/sig/internal/sigconfig:go_default_library",
         "//go/sig/internal/xnet:go_default_library",
-        "//go/sig/sigcmn:go_default_library",
         "@com_github_burntsushi_toml//:go_default_library",
         "@com_github_syndtr_gocapability//capability:go_default_library",
     ],
