+++ go/sig/egress/worker/BUILD.bazel
@@ -16,8 +16,8 @@ go_library(
         "//go/sig/egress/iface:go_default_library",
         "//go/sig/egress/siginfo:go_default_library",
         "//go/sig/internal/metrics:go_default_library",
+        "//go/sig/internal/sigcmn:go_default_library",
         "//go/sig/mgmt:go_default_library",
-        "//go/sig/sigcmn:go_default_library",
     ],
 )
