+++ go/sig/internal/base/BUILD.bazel
@@ -16,7 +16,7 @@ go_library(
         "//go/lib/log:go_default_library",
         "//go/lib/snet:go_default_library",
         "//go/sig/internal/disp:go_default_library",
+        "//go/sig/internal/sigcmn:go_default_library",
         "//go/sig/mgmt:go_default_library",
-        "//go/sig/sigcmn:go_default_library",
     ],
 )
