+++ go/sig/internal/ingress/BUILD.bazel
@@ -21,8 +21,8 @@ go_library(
         "//go/lib/sock/reliable:go_default_library",
         "//go/lib/util:go_default_library",
         "//go/sig/internal/metrics:go_default_library",
+        "//go/sig/internal/sigcmn:go_default_library",
         "//go/sig/mgmt:go_default_library",
-        "//go/sig/sigcmn:go_default_library",
     ],
 )
