+++ go/sig/internal/ingress/rlist.go
@@ -22,7 +22,7 @@ import (
 	"github.com/scionproto/scion/go/lib/common"
 	"github.com/scionproto/scion/go/lib/log"
 	"github.com/scionproto/scion/go/sig/internal/metrics"
-	"github.com/scionproto/scion/go/sig/sigcmn"
+	"github.com/scionproto/scion/go/sig/internal/sigcmn"
 )
 
 // ReassemblyList is used to keep a doubly linked list of SIG frames that are
