+++ go/sig/internal/sigcmn/common.go
